Splitted windbarb series into type files.
