+++ ts/Series/Windbarb/WindbarbPoint.ts
@@ -1,0 +1,62 @@
+/* *
+ *
+ *  Wind barb series module
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+  *
+  * Import
+  *
+  * */
+import type WindbarbPointOptions from './WindbarbPointOptions';
+import U from '../../Core/Utilities.js';
+const { isNumber } = U;
+import WindbarbSeries from './WindbarbSeries.js';
+import ColumnSeries from '../Column/ColumnSeries.js';
+
+/* *
+ *
+ * Class
+ *
+ * */
+
+class WindbarbPoint extends ColumnSeries.prototype.pointClass {
+
+    /* *
+     *
+     * Properties
+     *
+     * */
+    public beaufort: string = void 0 as any;
+    public beaufortLevel: number = void 0 as any;
+    public direction: number = void 0 as any;
+    public options: WindbarbPointOptions = void 0 as any;
+    public series: WindbarbSeries = void 0 as any;
+
+    /* *
+     *
+     * Functions
+     *
+     * */
+    public isValid(): boolean {
+        return isNumber(this.value) && this.value >= 0;
+    }
+}
+
+interface WindbarbPoint {
+    value: number;
+}
+
+/* *
+ *
+ * Default export
+ *
+ * */
+export default WindbarbPoint;
