+++ ts/Series/Windbarb/WindbarbPointOptions.d.ts
@@ -1,0 +1,35 @@
+/* *
+ *
+ *  Wind barb series module
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ * Imports
+ *
+ * */
+import type ColumnPointOptions from '../Column/ColumnPointOptions';
+
+/* *
+ *
+ * Declarations
+ *
+ * */
+export interface WindbarbPointOptions extends ColumnPointOptions {
+    direction?: number;
+    value?: number;
+}
+
+/* *
+ *
+ * Export default
+ *
+ * */
+export default WindbarbPointOptions;
